Add a link to the new playable simulation
The landing page for the site is going to be mentioned in my next video
as "the location" where demos and interactive simulations can be found.
So make sure the page now reflects this expectation.

This commit also provides a modified version of home.html in _layouts,
as the original one ends up showing a default title matching the first
heading in the page. This results in the title occurring twice.